+++ src/api/v2/index.ts
@@ -35,10 +35,12 @@ import { type Account, accounts, posts } from "../../schema";
 import { uuid } from "../../uuid";
 import { postMedia } from "../v1/media";
 import instance from "./instance";
+import notificationsRoutes from "./notifications";
 
 const app = new Hono<{ Variables: Variables }>();
 
 app.route("/instance", instance);
+app.route("/notifications", notificationsRoutes);
 
 app.post("/media", tokenRequired, scopeRequired(["write:media"]), postMedia);