+++ src/api/v2/instance.ts
@@ -30,6 +30,9 @@ app.get("/", async (c) => {
     )
     .groupBy(posts.language);
   return c.json({
+    api_versions: {
+      mastodon: 2,
+    },
     domain: url.host,
     title: url.host,
     version: metadata.version,